refactor(h-grid): Rename barrel file
